% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CramerLundberg.R
\name{CramerLundberg}
\alias{CramerLundberg}
\title{Constructs an object of CramerLundberg S4 class}
\usage{
CramerLundberg(initial_capital = NULL, premium_rate = NULL,
  claim_poisson_arrival_rate = NULL, claim_size_generator = NULL,
  claim_size_parameters = NULL)
}
\arguments{
\item{initial_capital}{a length one numeric non-negative vector specifying an
initial capital. Default: \code{0}.}

\item{premium_rate}{a length one numeric non-negative vector specifying a
premium rate. Default: \code{1}.}

\item{claim_poisson_arrival_rate}{a length one numeric positive vector
specifying the rate of the Poisson process of claims' arrivals. Default:
\code{1}.}

\item{claim_size_generator}{a function indicating the random generator of
claims' sizes. Default: \code{rexp}.}

\item{claim_size_parameters}{a named list containing parameters for the
random generator of claims' sizes. Default: \code{list(rate = 1)}.}
}
\value{
An object of \code{CramerLundberg} class.
}
\description{
\code{CramerLundberg()} constructs an object of \code{CramerLundberg} S4
class.
}
\examples{
model <- CramerLundberg(initial_capital = 10,
                        premium_rate = 1,
                        claim_poisson_arrival_rate = 1,
                        claim_size_generator = rexp,
                        claim_size_parameters = list(rate = 1))

}
